Basis compute shader mip map generation.
This method just uses the previously generated mip map to create a new mip map. The ideal solution would be to do what MiniEngine and other renderers do where  we take into account if dimensions are odd and even. This will have to be reviewed in the future.

+++ SandBox/SandBox.cpp
@@ -63,8 +63,8 @@ void SandBox::OnInit()
 
 		mScene->AddLight(mDevice.get(), pointLightcreationDesc2);
 
-		lightPos.x = (i) % 10;
-		lightPos.z = (i) / 10;
+		lightPos.x = static_cast<float>((i) % 10);
+		lightPos.z = static_cast<float>((i) / 10);
 
 		scene::Light::GetLightBufferData()->lightColor[i].x += lightPos.x / 255.0f;
 		scene::Light::GetLightBufferData()->lightColor[i].z += lightPos.z / 255.0f;
